scraper functionality working with bot command !
