+++ 896175.json
@@ -1,0 +1,9 @@
+{
+    "HLTV": "1.19",
+    "ADR": "96.3",
+    "HS": "0.105",
+    "W": "6",
+    "L": "5",
+    "T": "0",
+    "win_percent": ".545"
+}
