+++ commands/report.js
@@ -1,10 +1,11 @@
 const Discord = require("discord.js");
 
 module.exports.run = async (bot, message, args) => {
-
+    console.log(args);
     let rUser = message.guild.member(message.mentions.users.first() || message.guild.members.get(args[0]));
     if (!rUser) return message.channel.send("Couldn't find user");
     let reason = args.join(" ").slice(22);
+    
 
     let reportEmbed = new Discord.RichEmbed()
         .setDescription("Reports")
